added profileInfo
